Improve exception handling in apis/builders.clj
This update fine-tunes the exception handling within apis/builders.clj. We have made changes to catch AssertionError and general Exceptions separately during the execution of forms. In case of an AssertionError, the exception object is included in the thrown exception info map. Existing logging of the exceptions has been enhanced by extra data for better context.